--- .idea/inspectionProfiles/Project_Default.xml
@@ -1,13 +1,0 @@
-<component name="InspectionProjectProfileManager">
-  <profile version="1.0">
-    <option name="myName" value="Project Default" />
-    <inspection_tool class="PyPep8NamingInspection" enabled="true" level="WEAK WARNING" enabled_by_default="true">
-      <option name="ignoredErrors">
-        <list>
-          <option value="N802" />
-          <option value="N801" />
-        </list>
-      </option>
-    </inspection_tool>
-  </profile>
-</component>
